+++ packages/melonjs/src/application/defaultApplicationSettings.ts
@@ -9,6 +9,7 @@ export const defaultApplicationSettings = {
 	powerPreference: "default",
 	transparent: false,
 	antiAlias: false,
+	preWarmShaders: false,
 	textureFilter: "auto",
 	maxTextures: "auto",
 	castGroundShadow: true,